Added Energy Watcher API Added Event to Recalculate Channel Requirements for a Node. Level Emitter can now output redstone based on network energy levels. Level Emitter's entry field can now hold longs ( text box supports up o trillions )

+++ core/sync/packets/PacketValueConfig.java
@@ -55,7 +55,7 @@ public class PacketValueConfig extends AppEngPacket
 		else if ( Name.equals( "LevelEmitter.Value" ) && c instanceof ContainerLevelEmitter )
 		{
 			ContainerLevelEmitter lvc = (ContainerLevelEmitter) c;
-			lvc.setLevel( Integer.parseInt( Value ), player );
+			lvc.setLevel( Long.parseLong( Value ), player );
 			return;
 		}
 		else if ( Name.startsWith( "CellWorkbench." ) && c instanceof ContainerCellWorkbench )
